# server/apidocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beskar API",
    "version": "1.0.0",
    "description": "Authentication and implemented whiteboarding v2 APIs.\n\n**Try it:** [Sign in](/auth/login?returnTo=%2Fapi%2Fdocs), then use the browser-session option, or select BearerToken and paste a valid Zitadel JWT (without the Bearer prefix). The server currently verifies bearer JWTs using its OIDC ID-token verifier; opaque access tokens are not supported. Calls run against this origin with your permissions and can create real data.\n\nReplace example space/parent IDs with your own. Generate an Idempotency-Key UUID for each new board or checkpoint batch and keep it with the same payload for retries. Login, registration and logout require browser navigation. Cookie sessions normally require HTTPS.",
    "license": {
      "name": "Apache 2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This Beskar instance"
    }
  ],
...
  "security": [
    {
      "BrowserSession": []
    },
    {
      "BearerToken": []
    }
  ],
  "paths": {
...
    "/api/v2/editor/space/{spaceId}/whiteboard/create": {
      "post": {
        "tags": [
          "Whiteboarding v2"
        ],
        "operationId": "createWhiteboardV2",
        "summary": "Create a whiteboard",
        "description": "Creates a board in the v2 schema. Requires space edit_page permission and parent edit permission when nested. Generate a fresh UUID Idempotency-Key per create action; reuse it and the same body on retries. Replays return 201 with the original IDs. Unknown fields, invalid UTF-8 and trailing JSON are rejected. Maximum body: 16 KiB. V2 incremental checkpoints are supported. V2 retrieval and publication are not implemented yet; v1 readers cannot load this board.",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "example": "c7022348-1bb3-4e52-8403-17786e15e035"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "description": "Exactly one nonzero UUID. Generate with crypto.randomUUID(); retain for retries, replace for each new create action."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWhiteboardRequest"
              },
              "examples": {
                "topLevel": {
                  "summary": "Top-level board",
                  "value": {
                    "title": "Architecture discussion"
                  }
                },
                "nested": {
                  "summary": "Under an existing parent",
                  "value": {
                    "title": "Architecture discussion",
                    "parentId": 123
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created, or replayed successfully with the same page and space IDs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWhiteboardResponse"
                }
              }
            }
          },
          "302": {
            "description": "Browser redirect. Follow the Location header through the identity provider.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Redirect destination."
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED from the handler, or an authentication middleware error with a numeric code.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/WhiteboardError"
                    },
                    {
                      "$ref": "#/components/schemas/AuthenticationError"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid space UUID, Idempotency-Key, or JSON body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "INVALID_REQUEST",
                    "message": "Invalid space UUID, Idempotency-Key, or JSON body."
                  }
                }
              }
            }
          },
          "403": {
            "description": "Space edit_page or parent edit permission denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "CREATE_WHITEBOARD_FORBIDDEN",
                    "message": "Space edit_page or parent edit permission denied."
                  }
                }
              }
            }
          },
          "404": {
            "description": "Space or parent missing/deleted, or parent belongs to another space.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "SPACE_OR_PARENT_NOT_FOUND",
                    "message": "Space or parent missing/deleted, or parent belongs to another space."
                  }
                }
              }
            }
          },
          "409": {
            "description": "IDEMPOTENCY_KEY_REUSED: changed content with an existing key. SPACE_ARCHIVED: space is archived.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "IDEMPOTENCY_KEY_REUSED",
                    "message": "IDEMPOTENCY_KEY_REUSED: changed content with an existing key. SPACE_ARCHIVED: space is archived."
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "REQUEST_TOO_LARGE",
                    "message": "JSON body exceeds 16 KiB."
                  }
                }
              }
            }
          },
          "415": {
            "description": "Content-Type must be application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "UNSUPPORTED_MEDIA_TYPE",
                    "message": "Content-Type must be application/json."
                  }
                }
              }
            }
          },
          "500": {
            "description": "Creation could not be confirmed. Retry the same request and key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "WHITEBOARD_CREATE_FAILED",
                    "message": "Creation could not be confirmed. Retry the same request and key."
                  }
                }
              }
            }
          },
          "503": {
            "description": "WHITEBOARD_CREATE_BUSY: lock timeout. WHITEBOARD_PERMISSIONS_PENDING: creation recorded but permission setup pending. Retry the same request and key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "WHITEBOARD_CREATE_BUSY",
                    "message": "WHITEBOARD_CREATE_BUSY: lock timeout. WHITEBOARD_PERMISSIONS_PENDING: creation recorded but permission setup pending. Retry the same request and key."
                  }
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "1"
                  ]
                },
                "description": "Retry after one second."
              }
            }
          }
        }
      }
    },
    "/api/v2/editor/space/{spaceId}/whiteboard/{pageId}/checkpoint": {
      "post": {
        "tags": [
          "Whiteboarding v2"
        ],
        "operationId": "checkpointWhiteboardV2",
        "summary": "Save an incremental whiteboard checkpoint",
        "description": "Requires page edit permission and matching page/space membership. Appends one Yjs v1 batch and advances the existing draft atomically; the base snapshot remains unchanged. The draft row lock serializes sequence allocation. No expected-head or base-snapshot parameter is required.\n\nGenerate a UUID Idempotency-Key for each batch and retain it with the exact pending bytes until acknowledged. Matching retries return the original acknowledgement even if the head has advanced or the space was subsequently archived/soft-deleted, provided authorization and board membership still hold. Different actors may submit overlapping changes; Yjs handles their document effect. Acknowledgement confirms storage, not reception of other actors\u2019 edits.\n\nMaximum decoded update: 1 MiB (1048576 bytes). Maximum JSON body: 1.5 MiB (1572864 bytes). Client batching target: 256 KiB or two seconds of pending edits; that queue is not implemented by this endpoint. Images upload separately. Optional title metadata is saved atomically at this sequence. No snapshot compaction or permission provisioning occurs here. Unknown fields, extra JSON values and malformed Yjs v1 wire structures are rejected. Structural validation does not verify drawing schema or causal completeness. Deploy the whiteboard_updates and whiteboard_titles Liquibase changesets before using this operation. Title is authoritative checkpoint metadata; a title-only rename uses the empty Yjs update AAA=. Omitted title preserves the current value. Same key with another normalized title returns IDEMPOTENCY_KEY_REUSED.",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "example": "c7022348-1bb3-4e52-8403-17786e15e035"
          },
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "example": 42,
            "description": "Existing v2 whiteboard page ID in the supplied space."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "description": "Exactly one nonzero UUID per batch. Keep the same key and decoded bytes until acknowledged. Scoped by board and authenticated actor. Different content with the same scoped key returns 409."
          }
        ],
        "requestBody": {
          "required": true,
          "description": "One incremental batch, not the whole board snapshot.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointWhiteboardRequest"
              },
              "examples": {
                "emptyUpdate": {
                  "summary": "Valid empty Yjs v1 update (still records a checkpoint)",
                  "value": {
                    "updateEncoding": "yjs-update-v1",
                    "update": "AAA="
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch committed, or matching retry acknowledged with its original update ID and sequence.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointWhiteboardResponse"
                },
                "example": {
                  "status": "success",
                  "data": {
                    "pageId": 42,
                    "updateId": "0195ad23-831a-7000-8000-000000000100",
                    "sequence": "17"
                  }
                }
              }
            }
          },
          "302": {
            "description": "Browser redirect. Follow the Location header through the identity provider.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Redirect destination."
              }
            }
          },
          "400": {
            "description": "Invalid IDs, key, JSON, Base64, encoding, or Yjs v1 structure; nested lib0 values may not exceed 64 levels.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "INVALID_REQUEST",
                    "message": "Invalid Yjs v1 update or nesting exceeds 64 levels."
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED from the handler, or an authentication middleware error with a numeric code.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/WhiteboardError"
                    },
                    {
                      "$ref": "#/components/schemas/AuthenticationError"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Page edit permission denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "CHECKPOINT_FORBIDDEN",
                    "message": "You cannot edit this whiteboard."
                  }
                }
              }
            }
          },
          "404": {
            "description": "Board or space missing, board belongs to another space, or a new batch targets a deleted space.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "WHITEBOARD_NOT_FOUND",
                    "message": "Whiteboard not found in this space."
                  }
                }
              }
            }
          },
          "409": {
            "description": "Same scoped key with different content, or a new batch targets an archived space.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "examples": {
                  "keyReused": {
                    "value": {
                      "status": "FAILED",
                      "error": {
                        "code": "IDEMPOTENCY_KEY_REUSED",
                        "message": "idempotency key was reused with different content"
                      }
                    }
                  },
                  "spaceArchived": {
                    "value": {
                      "status": "FAILED",
                      "error": {
                        "code": "SPACE_ARCHIVED",
                        "message": "space is archived"
                      }
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds 1572864 bytes (1.5 MiB) or decoded update exceeds 1048576 bytes (1 MiB). Reduce the payload; do not retry unchanged.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "examples": {
                  "requestTooLarge": {
                    "value": {
                      "status": "FAILED",
                      "error": {
                        "code": "REQUEST_TOO_LARGE",
                        "message": "Checkpoint request exceeds 1.5 MiB."
                      }
                    }
                  },
                  "updateTooLarge": {
                    "value": {
                      "status": "FAILED",
                      "error": {
                        "code": "REQUEST_TOO_LARGE",
                        "message": "Decoded checkpoint exceeds 1 MiB."
                      }
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "Content type must be application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "UNSUPPORTED_MEDIA_TYPE",
                    "message": "Content-Type must be application/json."
                  }
                }
              }
            }
          },
          "500": {
            "description": "Save could not be confirmed, or draft integrity failed. Retry the same batch and key; persistent integrity failures need investigation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "WHITEBOARD_CHECKPOINT_FAILED",
                    "message": "Save could not be confirmed. Retry the same batch and Idempotency-Key."
                  }
                }
              }
            }
          },
          "503": {
            "description": "A database lock wait exceeded three seconds. Retry the same batch and key after one second. This is a per-lock timeout, not a total request deadline.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "WHITEBOARD_CHECKPOINT_BUSY",
                    "message": "Checkpoint is busy. Retry the same batch and Idempotency-Key."
                  }
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "1"
                  ]
                },
                "description": "Retry after one second."
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Valid Zitadel JWT accepted by the server OIDC verifier. Paste only the token."
      },
      "BrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zitadel.session",
        "description": "Sign in using the link above. The browser sends the HttpOnly session cookie automatically; do not paste a cookie value."
      }
    },
    "schemas": {
      "CreateWhiteboardRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Trimmed before validation; 1\u2013255 Unicode characters after trimming. NUL characters are rejected.",
            "example": "Architecture discussion"
          },
          "parentId": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Parent page in this space. Omit or use null for a top-level board."
          }
        }
      },
      "CreateWhiteboardResponse": {
        "type": "object",
        "required": [
          "status",
          "data"
        ],
        "properties": {
          "status": {
            "const": "success"
          },
          "data": {
            "type": "object",
            "required": [
              "pageId",
              "spaceId"
            ],
            "properties": {
              "pageId": {
                "type": "integer",
                "format": "int64",
                "example": 42
              },
              "spaceId": {
                "type": "string",
                "format": "uuid",
                "description": "Nonzero UUID.",
                "example": "c7022348-1bb3-4e52-8403-17786e15e035"
              }
            }
          }
        }
      },
      "WhiteboardError": {
        "type": "object",
        "required": [
          "status",
          "error"
        ],
        "properties": {
          "status": {
            "const": "FAILED"
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "AuthenticationError": {
        "type": "object",
        "required": [
          "status",
          "error"
        ],
        "properties": {
          "status": {
            "const": "FAILED"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "example": 401
              },
              "message": {
                "type": "string"
              },
              "detail": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "CheckpointWhiteboardRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "updateEncoding",
          "update"
        ],
        "properties": {
          "updateEncoding": {
            "type": "string",
            "enum": [
              "yjs-update-v1"
            ],
            "description": "Yjs binary update format; maps to update_encoding. Independent of the API version."
          },
          "update": {
            "type": "string",
            "contentEncoding": "base64",
            "minLength": 1,
            "description": "Standard Base64 encoding of a structurally valid Yjs v1 update. Decoded bytes must be nonempty and at most 1048576 bytes. The entire JSON body must fit within 1572864 bytes. Do not split binary updates at arbitrary byte offsets.",
            "example": "AAA="
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Optional rename at this checkpoint sequence. Trimmed; null/blank/NUL rejected. Use update AAA= for a title-only rename. Omit to preserve title."
          },
          "restoreGeneration": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$",
            "description": "Draft restore generation from the manifest. Omission means 0; stale generations receive 409 WHITEBOARD_RESTORED.",
            "default": "0"
          }
        }
      },
      "CheckpointWhiteboardResponse": {
        "type": "object",
        "required": [
          "status",
          "data"
        ],
        "properties": {
          "status": {
            "const": "success"
          },
          "data": {
            "type": "object",
            "required": [
              "pageId",
              "updateId",
              "sequence"
            ],
            "properties": {
              "pageId": {
                "type": "integer",
                "format": "int64",
                "minimum": 1,
                "maximum": 9007199254740991,
                "example": 42
              },
              "updateId": {
                "type": "string",
                "format": "uuid",
                "description": "Server-generated batch identity; stable across matching retries."
              },
              "sequence": {
                "type": "string",
                "pattern": "^[1-9][0-9]{0,18}$",
                "description": "This batch\u2019s committed position, from 1 to 9223372036854775807. Decimal string preserves BIGINT precision. Replays return the original position, not the current draft head.",
                "example": "17"
              }
            }
          }
        }
      },
...
    }
  }
}